+++ tests/persist-test.c
@@ -1,4 +1,4 @@
-/*
+/* tests/persist-test.c
  * Copyright (c) 2000-2004
  *      The Regents of the University of California.  All rights reserved.
  *
@@ -48,7 +48,7 @@
 
 /* Types */
 struct node_ {
-  int data;
+  intptr_t data;
   struct node_ *left;
   struct node_ *right;
 };
@@ -62,7 +62,7 @@ static region test_rgn;
 bool node_serialize(FILE *f, void *obj)
 {
   node n = (node) obj;
-  
+
   assert(f);
   assert(obj);
 
@@ -80,7 +80,7 @@ bool node_serialize(FILE *f, void *obj)
 void *node_deserialize(FILE *f)
 {
   node n = ralloc(test_rgn, struct node_);
-  
+
   fread((void *)&n->data, sizeof(int), 1, f);
   fread((void *)&n->left, sizeof(void *), 1, f);
   fread((void *)&n->right, sizeof(void *), 1, f);
@@ -110,31 +110,31 @@ static void pt_serialize()
   node n1;
   hash_table table;
   /* Build a small graph and a hash table */
-  { 				
+  {
     node n2,n3,n4;
-    
-    table = make_persistent_string_hash_table(4, 
+
+    table = make_persistent_string_hash_table(4,
 					      0);
 
     n1 = ralloc(test_rgn, struct node_);
     n2 = ralloc(test_rgn, struct node_);
     n3 = ralloc(test_rgn, struct node_);
     n4 = ralloc(test_rgn, struct node_);
-    
+
     n1->data = 1;
     n2->data = 2;
     n3->data = 3;
     n4->data = 4;
-    
+
     n1->left = n2;
     n1->right = n3;
-    
+
     n2->left = n4;
     n2->right = n4;
-    
+
     n3->left = n4;
     n3->right = n4;
-    
+
     n4->left = n1;
     n4->right = n1;
 
@@ -145,10 +145,10 @@ static void pt_serialize()
   }
 
   /* Serialize it */
-  {  
+  {
     FILE *outfile;
-    serialize_fn_ptr serialize_fns[5] = {nonptr_data_serialize, 
-					 funptr_data_serialize, 
+    serialize_fn_ptr serialize_fns[5] = {nonptr_data_serialize,
+					 funptr_data_serialize,
 					 string_data_serialize,
 					 node_serialize,
 					 hash_table_serialize};
@@ -157,10 +157,10 @@ static void pt_serialize()
 
     /* Write down the old address of n1 */
     fwrite((void *)&n1, sizeof(void *), 1, outfile);
-    
+
     /* Write down the old address of table */
     fwrite((void *)&table, sizeof(void *), 1, outfile);
-    
+
     serialize_start(outfile, serialize_fns, 5);
     serialize_object(n1, NODE_PERSIST_KIND);
     serialize_object(table,HASH_PERSIST_KIND);
@@ -187,7 +187,7 @@ static void pt_deserialize()
 					 hash_table_set_fields};
 
   infile = fopen(DEFAULT_FILENAME, "rb");
-  
+
   if (infile == NULL) {
     printf("Couldn't open %s\n", DEFAULT_FILENAME);
     exit(1);
@@ -200,7 +200,7 @@ static void pt_deserialize()
   fread((void *)&table, sizeof(void *), 1, infile);
 
   deserialize_all(infile, deserialize_fns, set_fields_fns, 5);
-  
+
   /* Now update n with its new address */
   deserialize_set_obj((void **)&n);
 
@@ -225,6 +225,7 @@ static void pt_deserialize()
   if (n->left->right != n->left->left) {
     fail("n2's children were not deserialized correctly.\n");
   }
+  /* FIXME: -Wtautological-compare: */
   if (n->right->left != n->right->left) {
     fail("n3's children were not deserialized correctly.\n");
   }
@@ -237,7 +238,7 @@ static void pt_deserialize()
   if (n->left->left->left != n || n->left->left->right != n) {
     fail("n4's children were not deserialized correctly.\n");
   }
-  
+
   /* Check well-formedness of the hash table */
   {
     hash_table_scanner scan;
@@ -247,22 +248,22 @@ static void pt_deserialize()
     hash_table_scan(table,&scan);
 
     while(hash_table_next(&scan, &next_key, &next_data)) {
-      if ((int)next_data == 1) {
+      if ((intptr_t)next_data == 1) {
 	if (strcmp("n1",next_key)) {
 	  fail("hash table doesn't map n1 correctly.\n");
 	}
       }
-      else if ((int)next_data == 2) {
+      else if ((intptr_t)next_data == 2) {
 	if (strcmp("n2",next_key)) {
 	  fail("hash table doesn't map n2 correctly.\n");
 	}
       }
-      else if ((int)next_data == 3) {
+      else if ((intptr_t)next_data == 3) {
 	if (strcmp("n3",next_key)) {
 	  fail("hash table doesn't map n3 correctly.\n");
 	}
       }
-      else if ((int)next_data == 4) {
+      else if ((intptr_t)next_data == 4) {
 	if (strcmp("n4",next_key)) {
 	  fail("hash toble doesn't map n4 correctly.\n");
 	}
@@ -271,14 +272,14 @@ static void pt_deserialize()
 	fail("There is an unknown mapping in the hash table.\n");
       }
     }
-    
+
   }
 }
 
 void seed_fn_ptr_table(region r);
 
 int main(int argc, char **argv)
-{ 
+{
   assert(argc > 0);
 
   region_init();
